requirements.txt
holidays==0.53
langchain==0.2.11
langchain_openai==0.1.17
matplotlib==3.9.1
openai==1.37.0
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.37.0
tabulate==0.9.0
urllib3==1.26.15